+++ sdk-node/test/conformance/runner.test.ts
@@ -14,6 +14,7 @@ import { setupMockServer, TEST_BASE_URL } from '../helpers/mock-server'
 import { dispatchContacts } from './dispatch-contacts'
 import { dispatchEntitlements } from './dispatch-entitlements'
 import { dispatchEvents } from './dispatch-events'
+import { dispatchFeatures } from './dispatch-features'
 import { dispatchInvoices } from './dispatch-invoices'
 import { dispatchPrices } from './dispatch-prices'
 import { dispatchSubscriptions } from './dispatch-subscriptions'
@@ -102,6 +103,8 @@ function dispatchCall(
       return dispatchEntitlements(client, call)
     case 'prices':
       return dispatchPrices(client, call)
+    case 'features':
+      return dispatchFeatures(client, call)
   }
 }